Change the rootlogger setup
Also respect the singleton pattern for the global settings object.

This was required, as the basicConfig call actually is not respected if a call
has been made to log previously. Then the logging package has already done it's
basic configuration, and the call will be ignored.

This tripped me up for quite some time, and ideally this should be refactored
slightly later.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>