MEN-4434: Report a failed update in case of missing 'install' script
Previously, an error with the install script - either a permission issue, or the
script simply missing would not cause an error log to be updated to the Mender
server upon failure.

This PR amends this, and uploads the deployment log in both cases

Changelog: Report a failed update in case of a issues with the `install` script.
In case of a missing script, or a permissions issue, the deployment log is
updated to the Mender server before the client exits as before.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>